添加jmx数据获取的处理

+++ machairodus-balancer/src/main/java/org/machairodus/balancer/component/impl/CommandComponentImpl.java
@@ -17,6 +17,7 @@ package org.machairodus.balancer.component.impl;
 
 import org.machairodus.balancer.component.CommandComponent;
 import org.machairodus.balancer.service.CommandService;
+import org.machairodus.commons.token.Token;
 import org.machairodus.commons.util.ResponseStatus;
 import org.machairodus.mappers.domain.NodeConfig;
 import org.nanoframework.core.status.ResultMap;
@@ -29,15 +30,22 @@ public class CommandComponentImpl implements CommandComponent {
 	private CommandService commandService;
 	
 	@Override
-	public Object create(NodeConfig node) {
+	public Object create(NodeConfig node, String token) {
 		if(node.getId() == null) 
 			return ResultMap.create(ResponseStatus.FAIL.getStatus(), "无效的节点配置对象: ID不能为空", ResponseStatus.FAIL.getInfo());
+		
+		if(!Token.decode(token)) 
+			return ResultMap.create(ResponseStatus.FAIL.getStatus(), "无效的Token认证", ResponseStatus.FAIL.getInfo());
+		
 		
 		return commandService.createScheduler(node);
 	}
 
 	@Override
-	public Object destroy(Long id) {
+	public Object destroy(Long id, String token) {
+		if(!Token.decode(token)) 
+			return ResultMap.create(ResponseStatus.FAIL.getStatus(), "无效的Token认证", ResponseStatus.FAIL.getInfo());
+		
 		return commandService.destroyScheduler(id);
 	}
